+++ RELEASES.md
@@ -7,3 +7,11 @@ Based on:
 - Speakeasy CLI 1.299.3 (2.338.7) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [typescript v0.0.3] ./packages/client-ledger
+
+## 2024-06-04 00:15:11
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.299.6 (2.338.12) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.0.4] ./packages/client-payments